tests: Allow insecure ssh key exchange
The key exchange algorithms supported by sshd in the docker images used
for integration tests are considered insecure by modern versions of ssh.
This causes the ssh connection to fail when running tests on Ubuntu
20.04.

Since these connections are to ephemeral Docker containers, spawned by
the test suite, on localhost I believe it safe to enable one of the
insecure algorithms.